+++ packaging/chocolatey/0.2.1/tools/chocolateybeforemodify.ps1
@@ -1,0 +1,2 @@
+﻿$ErrorActionPreference = 'Stop'
+Uninstall-BinFile -Name 'cliai'